Conformance: POOL: Translate and adjust.
- Refactor modules to have a Base and per-rule ones.
- Confirm the translation of data-types for POOL.
- Adjust Specification for PParams.
- Adjust Specification for PoolTxCert to omit RetirePool with a note:
since the spec does not yet implement the two predicate failures from
Shelley.
